refactor(trans.content): 重构翻译组件结构并迁移至 src/components将原 TransLine 组件及相关逻辑从 entrypoints/trans.content/script 目录迁移至src/components/transline，并拆分为多个模块化组件，提升代码可维护性和复用性。
- 移除旧 TransLine.tsx 文件
- 新增 TransLine、T2、HoverTooltip、PanelComponents 和 transUtils 等组件
- 优化 DOM 操作工具函数导入路径- 为翻译功能相关核心逻辑建立统一组件目录结构
```
