+++ src/components/transline/index.ts
@@ -1,0 +1,1 @@
+export { default as TransLine } from './TransLine';
